@@ -11,16 +11,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^20.0.4",
-    "@angular/cdk": "^20.0.3",
-    "@angular/common": "^20.0.4",
-    "@angular/compiler": "^20.0.4",
-    "@angular/core": "^20.0.4",
-    "@angular/forms": "^20.0.4",
-    "@angular/platform-browser": "^20.0.4",
-    "@angular/platform-browser-dynamic": "^20.0.4",
-    "@angular/router": "^20.0.4",
+    "@angular/animations": "20.1.2",
+    "@angular/cdk": "20.1.2",
+    "@angular/common": "20.1.2",
+    "@angular/compiler": "20.1.2",
+    "@angular/core": "20.1.2",
+    "@angular/forms": "20.1.2",
+    "@angular/localize": "20.1.2",
+    "@angular/platform-browser": "20.1.2",
+    "@angular/platform-browser-dynamic": "20.1.2",
+    "@angular/router": "20.1.2",
     "@ng-select/ng-select": "^15.0.0",
+    "@ngx-translate/core": "^16.0.4",
+    "@ngx-translate/http-loader": "^16.0.1",
     "ag-grid-angular": "^33.3.2",
     "ag-grid-community": "^33.3.2",
     "apexcharts": "^4.7.0",
